requirements.txt
asgiref==3.8.1
azure-core==1.30.1
azure-storage-blob==12.20.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
fuzzywuzzy==0.6.1
idna==3.7
isodate==0.6.1
Levenshtein==0.25.1
mssql-django==1.5
pillow==10.3.0
psycopg==3.1.18
psycopg2==2.9.9
pycparser==2.22
pyodbc==5.1.0
python-decouple==3.8
python-Levenshtein==0.25.1
pytz==2024.1
rapidfuzz==3.9.0
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
